+++ packages/platforms/browser/lib/browser.ts
@@ -1,15 +1,14 @@
-import { createClient } from '@bugsnag/core-performance'
+import { createClient, InMemoryQueue } from '@bugsnag/core-performance'
+import createFetchDeliveryFactory from '@bugsnag/delivery-fetch-performance'
+import { createFetchRequestTracker, createXmlHttpRequestTracker } from '@bugsnag/request-tracker-performance'
 import { FullPageLoadPlugin, NetworkRequestPlugin, ResourceLoadPlugin, RouteChangePlugin } from './auto-instrumentation'
 import createBrowserBackgroundingListener from './backgrounding-listener'
 import createClock from './clock'
 import { createSchema } from './config'
 import { createDefaultRoutingProvider } from './default-routing-provider'
-import createBrowserDeliveryFactory from './delivery'
 import idGenerator from './id-generator'
 import createOnSettle from './on-settle'
 import makeBrowserPersistence from './persistence'
-import createFetchRequestTracker from './request-tracker/request-tracker-fetch'
-import createXmlHttpRequestTracker from './request-tracker/request-tracker-xhr'
 import createResourceAttributesSource from './resource-attributes-source'
 import createSpanAttributesSource from './span-attributes-source'
 import { WebVitals } from './web-vitals'
@@ -20,7 +19,7 @@ const clock = createClock(performance, backgroundingListener)
 const persistence = makeBrowserPersistence(window)
 const resourceAttributesSource = createResourceAttributesSource(navigator, persistence)
 const fetchRequestTracker = createFetchRequestTracker(window, clock)
-const xhrRequestTracker = createXmlHttpRequestTracker(window, clock)
+const xhrRequestTracker = createXmlHttpRequestTracker(XMLHttpRequest, clock, document)
 const webVitals = new WebVitals(performance, clock, window.PerformanceObserver)
 export const onSettle = createOnSettle(
   clock,
@@ -36,7 +35,7 @@ const BugsnagPerformance = createClient({
   clock,
   resourceAttributesSource,
   spanAttributesSource,
-  deliveryFactory: createBrowserDeliveryFactory(window.fetch, backgroundingListener, clock),
+  deliveryFactory: createFetchDeliveryFactory(window.fetch, clock, backgroundingListener),
   idGenerator,
   schema: createSchema(window.location.hostname, new DefaultRoutingProvider()),
   plugins: (spanFactory, spanContextStorage) => [
@@ -56,7 +55,8 @@ const BugsnagPerformance = createClient({
     new NetworkRequestPlugin(spanFactory, fetchRequestTracker, xhrRequestTracker),
     new RouteChangePlugin(spanFactory, window.location, document)
   ],
-  persistence
+  persistence,
+  retryQueueFactory: (delivery, retryQueueMaxSize) => new InMemoryQueue(delivery, retryQueueMaxSize)
 })
 
 export default BugsnagPerformance