+++ packages/platforms/react-native/lib/auto-instrumentation/index.ts
@@ -1,0 +1,2 @@
+export * from './app-start-plugin'
+export * from './network-request-plugin'