+++ packages/platforms/react-native/lib/index.ts
@@ -1,0 +1,3 @@
+import BugsnagPerformance from './client'
+
+export default BugsnagPerformance